rustfix
Function get_suggestions_from_json

pub fn get_suggestions_from_json<S: BuildHasher>(
    input: &str,
    only: &HashSet<String, S>,
    filter: Filter,
) -> Result<Vec<Suggestion>>

Collects code Suggestions from one or more compiler diagnostic lines.

Fails if any of diagnostic line input is not a valid Diagnostic JSON.

only — only diagnostics with code in a set of error codes would be collected.